Replace pulsing colors by progress stripes when building
Previously a stage was pulsing with fixed colors while building. This
hid the outcome of the last build. It was not visible whether the build
had been successful or not.

Now the stage is overlaid with gray progress stripes while building. The
outcome of the last build is still clearly visible (green or red).

Furthermore the CSS effects are now working on IE (prefixed `-ms`) and
standards-compliant browsers as well (without prefix).

NB: some statements are currently duplicated for each browser specific
prefix. There might be a SASS way of reducing this duplication.